+++ crates/core/Cargo.toml
@@ -26,9 +26,23 @@ serde_json.workspace = true
 regex.workspace = true
 path-slash.workspace = true
 alloy-primitives.workspace = true
+once_cell.workspace = true
+svm = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true }
 
 walkdir = "2.4"
 dunce = "1.0"
 memmap2 = "0.9"
 cfg-if = "1.0.0"
-once_cell = "1.19"
+
+tempfile = { version = "3.9", optional = true }
+fs_extra = { version = "1.3", optional = true }
+
+[dev_dependencies]
+solang-parser.workspace = true
+
+[features]
+project-util = ["dep:tempfile", "dep:fs_extra"]
+svm-solc = ["dep:svm", "dep:tokio"]
+async = ["dep:tokio"]
+test-utils = []
