+++ crates/core/src/lib.rs
@@ -1,2 +1,4 @@
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
 pub mod error;
 pub mod utils;
