+++ crates/project/src/compile/project.rs
@@ -625,12 +625,14 @@ fn compile_parallel<C: Compiler>(
 mod tests {
     use std::path::Path;
 
-    use super::*;
+    use foundry_compilers_artifacts::output_selection::ContractOutputSelection;
+
     use crate::{
-        artifacts::output_selection::ContractOutputSelection, compilers::multi::MultiCompiler,
-        project_util::TempProject, ConfigurableArtifacts, MinimalCombinedArtifacts,
-        ProjectPathsConfig,
+        compilers::multi::MultiCompiler, project_util::TempProject, ConfigurableArtifacts,
+        MinimalCombinedArtifacts, ProjectPathsConfig,
     };
+
+    use super::*;
 
     fn init_tracing() {
         let _ = tracing_subscriber::fmt()
@@ -641,7 +643,7 @@ mod tests {
 
     #[test]
     fn can_preprocess() {
-        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-data/dapp-sample");
+        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../test-data/dapp-sample");
         let project = Project::builder()
             .paths(ProjectPathsConfig::dapptools(root).unwrap())
             .build(Default::default())
@@ -660,7 +662,7 @@ mod tests {
 
     #[test]
     fn can_detect_cached_files() {
-        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-data/dapp-sample");
+        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../test-data/dapp-sample");
         let paths = ProjectPathsConfig::builder().sources(root.join("src")).lib(root.join("lib"));
         let project = TempProject::<MultiCompiler, MinimalCombinedArtifacts>::new(paths).unwrap();
 
@@ -795,7 +797,7 @@ mod tests {
 
     #[test]
     fn extra_output_cached() {
-        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-data/dapp-sample");
+        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../test-data/dapp-sample");
         let paths = ProjectPathsConfig::builder().sources(root.join("src")).lib(root.join("lib"));
         let mut project = TempProject::<MultiCompiler>::new(paths.clone()).unwrap();
