+++ crates/project/src/compilers/mod.rs
@@ -21,6 +21,7 @@ use std::{
 pub mod multi;
 pub mod solc;
 pub mod vyper;
+pub use vyper::*;
 
 /// A compiler version is either installed (available locally) or can be downloaded, from the remote
 /// endpoint
